Update R model.R
This was the initial code for our group work, starting with a simple manipulation of the data and then using some of the functions in ggplot2 and GGally, such as ggplot, ggparis, to draw some box plots, histograms etc. We observed that the relationship between the response variable and the independent variable was not linear and decided to use Poisson regression.  
Finally we used the glm function and filtered the variables according to significance levels to obtain the optimal model.